// src/app/models/user.model.ts
export interface User {
    uid: string;
    userName: string;
    userFirstName: string;
    userLastName: string;
    userPassword: string;
    userdesignation: string;
    websiteuser: string;
    userphone: string;
    usercountry: string;
    userstate: string;
    usercity: string;
    verified: boolean;
    summary: string;
    userlinkden: string;
    usergithub: string;
    otherturluser: string;
    accdeactivate: boolean;
  }
  